# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/health": {
      "get": {
        "operationId": "HealthController_check",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Vérifier que l’API est disponible",
        "tags": [
          "health"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Connexion réussie."
          },
          "401": {
            "description": "Identifiants invalides."
          }
        },
        "summary": "Connexion utilisateur et génération de JWT",
        "tags": [
          "auth"
        ]
      }
    },
    "/users": {
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Créer un utilisateur adhérent",
        "tags": [
          "users"
        ]
      },
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Lister les utilisateurs",
        "tags": [
          "users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Récupérer un utilisateur par son identifiant",
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Mettre à jour un utilisateur",
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Supprimer un utilisateur",
        "tags": [
          "users"
        ]
      }
    },
    "/profiles": {
      "post": {
        "operationId": "ProfilesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProfileDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Créer un profil adhérent (One-to-One)",
        "tags": [
          "profiles"
        ]
      },
      "get": {
        "operationId": "ProfilesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Lister les profils",
        "tags": [
          "profiles"
        ]
      }
    },
    "/profiles/{id}": {
      "get": {
        "operationId": "ProfilesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Récupérer un profil",
        "tags": [
          "profiles"
        ]
      },
      "patch": {
        "operationId": "ProfilesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Mettre à jour un profil",
        "tags": [
          "profiles"
        ]
      },
      "delete": {
        "operationId": "ProfilesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Supprimer un profil",
        "tags": [
          "profiles"
        ]
      }
    },
    "/authors": {
      "post": {
        "operationId": "AuthorsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAuthorDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Créer un auteur",
        "tags": [
          "authors"
        ]
      },
      "get": {
        "operationId": "AuthorsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Lister les auteurs",
        "tags": [
          "authors"
        ]
      }
    },
    "/authors/{id}": {
      "get": {
        "operationId": "AuthorsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Récupérer un auteur",
        "tags": [
          "authors"
        ]
      },
      "patch": {
        "operationId": "AuthorsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAuthorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Mettre à jour un auteur",
        "tags": [
          "authors"
        ]
      },
      "delete": {
        "operationId": "AuthorsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Supprimer un auteur",
        "tags": [
          "authors"
        ]
      }
    },
    "/categories": {
      "post": {
        "operationId": "CategoriesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Créer une catégorie",
        "tags": [
          "categories"
        ]
      },
      "get": {
        "operationId": "CategoriesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Lister les catégories",
        "tags": [
          "categories"
        ]
      }
    },
    "/categories/{id}": {
      "get": {
        "operationId": "CategoriesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Récupérer une catégorie",
        "tags": [
          "categories"
        ]
      },
      "patch": {
        "operationId": "CategoriesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Mettre à jour une catégorie",
        "tags": [
          "categories"
        ]
      },
      "delete": {
        "operationId": "CategoriesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Supprimer une catégorie",
        "tags": [
          "categories"
        ]
      }
    },
    "/books": {
      "post": {
        "operationId": "BooksController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Créer un livre avec sa catégorie et ses auteurs (Many-to-Many)",
        "tags": [
          "books"
        ]
      },
      "get": {
        "operationId": "BooksController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Lister les livres",
        "tags": [
          "books"
        ]
      }
    },
    "/books/{id}": {
      "get": {
        "operationId": "BooksController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Récupérer un livre",
        "tags": [
          "books"
        ]
      },
      "patch": {
        "operationId": "BooksController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBookDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Mettre à jour un livre",
        "tags": [
          "books"
        ]
      },
      "delete": {
        "operationId": "BooksController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Supprimer un livre",
        "tags": [
          "books"
        ]
      }
    },
    "/loans": {
      "post": {
        "operationId": "LoansController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLoanDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Créer un emprunt de livre",
        "tags": [
          "loans"
        ]
      },
      "get": {
        "operationId": "LoansController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Lister les emprunts",
        "tags": [
          "loans"
        ]
      }
    },
    "/loans/{id}": {
      "get": {
        "operationId": "LoansController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Récupérer un emprunt",
        "tags": [
          "loans"
        ]
      },
      "patch": {
        "operationId": "LoansController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLoanDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Mettre à jour un emprunt (ex: retour du livre)",
        "tags": [
          "loans"
        ]
      },
      "delete": {
        "operationId": "LoansController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Supprimer un emprunt",
        "tags": [
          "loans"
        ]
      }
    }
  },
  "info": {
    "title": "API Bibliothèque",
    "description": "API RESTful de gestion de bibliothèque réalisée avec NestJS, Sequelize et MySQL.",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "health",
      "description": ""
    },
    {
      "name": "users",
      "description": ""
    },
    {
      "name": "profiles",
      "description": ""
    },
    {
      "name": "authors",
      "description": ""
    },
    {
      "name": "categories",
      "description": ""
    },
    {
      "name": "books",
      "description": ""
    },
    {
      "name": "loans",
      "description": ""
    }
  ],
  "servers": [],
  "components": {
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "admin@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "admin123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "Amina"
          },
          "lastName": {
            "type": "string",
            "example": "Bensaid"
          },
          "email": {
            "type": "string",
            "example": "amina@example.com"
          },
          "address": {
            "type": "string",
            "example": "12 rue des Fleurs, Paris"
          },
          "phone": {
            "type": "string",
            "example": "+33601020304"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "email",
          "address",
          "phone"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "Amina"
          },
          "lastName": {
            "type": "string",
            "example": "Bensaid"
          },
          "email": {
            "type": "string",
            "example": "amina@example.com"
          },
          "address": {
            "type": "string",
            "example": "12 rue des Fleurs, Paris"
          },
          "phone": {
            "type": "string",
            "example": "+33601020304"
          }
        }
      },
      "CreateProfileDto": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string",
            "example": "12 rue des Fleurs, Paris"
          },
          "phone": {
            "type": "string",
            "example": "+33601020304"
          },
          "membershipNumber": {
            "type": "string",
            "example": "MEM-2026-001"
          },
          "joinedAt": {
            "type": "string",
            "example": "2026-04-15T10:00:00.000Z"
          },
          "userId": {
            "type": "number",
            "example": 1
          }
        },
        "required": [
          "address",
          "phone",
          "membershipNumber",
          "joinedAt",
          "userId"
        ]
      },
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string",
            "example": "12 rue des Fleurs, Paris"
          },
          "phone": {
            "type": "string",
            "example": "+33601020304"
          },
          "membershipNumber": {
            "type": "string",
            "example": "MEM-2026-001"
          },
          "joinedAt": {
            "type": "string",
            "example": "2026-04-15T10:00:00.000Z"
          },
          "userId": {
            "type": "number",
            "example": 1
          }
        }
      },
      "CreateAuthorDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "Victor"
          },
          "lastName": {
            "type": "string",
            "example": "Hugo"
          },
          "bio": {
            "type": "string",
            "example": "Écrivain français majeur du XIXe siècle."
          }
        },
        "required": [
          "firstName",
          "lastName"
        ]
      },
      "UpdateAuthorDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "Victor"
          },
          "lastName": {
            "type": "string",
            "example": "Hugo"
          },
          "bio": {
            "type": "string",
            "example": "Écrivain français majeur du XIXe siècle."
          }
        }
      },
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Roman"
          },
          "description": {
            "type": "string",
            "example": "Ouvrages de fiction littéraire."
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Roman"
          },
          "description": {
            "type": "string",
            "example": "Ouvrages de fiction littéraire."
          }
        }
      },
      "CreateBookDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Les Misérables"
          },
          "isbn": {
            "type": "string",
            "example": "978-2070409189"
          },
          "publishedYear": {
            "type": "number",
            "example": 1862
          },
          "totalCopies": {
            "type": "number",
            "example": 5
          },
          "availableCopies": {
            "type": "number",
            "example": 5
          },
          "summary": {
            "type": "string",
            "example": "Roman historique et social."
          },
          "coverUrl": {
            "type": "string",
            "example": "https://example.com/cover.jpg"
          },
          "categoryId": {
            "type": "number",
            "example": 1
          },
          "authorIds": {
            "example": [
              1,
              2
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "title",
          "isbn",
          "publishedYear",
          "totalCopies",
          "availableCopies",
          "categoryId",
          "authorIds"
        ]
      },
      "UpdateBookDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Les Misérables"
          },
          "isbn": {
            "type": "string",
            "example": "978-2070409189"
          },
          "publishedYear": {
            "type": "number",
            "example": 1862
          },
          "totalCopies": {
            "type": "number",
            "example": 5
          },
          "availableCopies": {
            "type": "number",
            "example": 5
          },
          "summary": {
            "type": "string",
            "example": "Roman historique et social."
          },
          "coverUrl": {
            "type": "string",
            "example": "https://example.com/cover.jpg"
          },
          "categoryId": {
            "type": "number",
            "example": 1
          },
          "authorIds": {
            "example": [
              1,
              2
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "CreateLoanDto": {
        "type": "object",
        "properties": {
          "dueDate": {
            "type": "string",
            "example": "2026-04-20T10:00:00.000Z"
          },
          "userId": {
            "type": "number",
            "example": 1
          },
          "bookId": {
            "type": "number",
            "example": 1
          }
        },
        "required": [
          "dueDate",
          "userId",
          "bookId"
        ]
      },
      "UpdateLoanDto": {
        "type": "object",
        "properties": {
          "dueDate": {
            "type": "string",
            "example": "2026-04-20T10:00:00.000Z"
          },
          "userId": {
            "type": "number",
            "example": 1
          },
          "bookId": {
            "type": "number",
            "example": 1
          },
          "returnedAt": {
            "type": "string",
            "example": "2026-04-18T14:00:00.000Z"
          },
          "status": {
            "type": "string",
            "example": "RETURNED",
            "enum": [
              "BORROWED",
              "RETURNED"
            ]
          }
        }
      }
    }
  }
}